sede attributes modularized and compiles.

+++ core/src/main/java/com/pulapirata/core/utils/PetAttributes.java
@@ -49,16 +49,16 @@ public class PetAttributes {
   private PetAttribute disciplina_;
   public PetAttribute disciplina() { return disciplina_; }
 
-  public PetAttributes(float beatsCoelhoHora) {
+  public PetAttributes(double beatsCoelhoHora) {
     alcool_   = new PetAttribute("Alcool", 3, 0, 10, -1, (int)beatsCoelhoHora);
     fome_     = new PetAttribute("Fome", 13, -20, 120, 10, (int)beatsCoelhoHora);
-    humor_    = new PetAttribute("Humor", 30, -20, 120, -5, (int)beatsCoelhoHora/3.);
+    humor_    = new PetAttribute("Humor", 30, -20, 120, -5, (int)(beatsCoelhoHora/3.));
     sede_     = new PetAttribute("Sede", 5, 0, 10, 1, (int)beatsCoelhoHora);
-    social_   = new PetAttribute("Social", 30, -20, 120, -5, (int)beatsCoelhoHora*2./3.);
-    higiene_  = new PetAttribute("Higiene", 30, -20, 120, -5, (int)beatsCoelhoHora/2.);
-    estudo_   = new PetAttribute("Estudo", 0, -1, -5, 10, (int)beatsCoelhoHora/24.);
+    social_   = new PetAttribute("Social", 30, -20, 120, -5, (int)(beatsCoelhoHora*2./3.));
+    higiene_  = new PetAttribute("Higiene", 30, -20, 120, -5, (int)(beatsCoelhoHora/2.));
+    estudo_   = new PetAttribute("Estudo", 0, -1, -5, 10, (int)(beatsCoelhoHora/24.));
       //? por dia a partir da matricula (colocar um valor inicial depois da matricula mudar)
-    saude_ = new PetAttribute("Saude", 5, -5, 10, -1, (int)beatsCoelhoHora*24.);
+    saude_ = new PetAttribute("Saude", 5, -5, 10, -1, (int)(beatsCoelhoHora*24.));
       //? por idade (em dias?)
     disciplina_ = new PetAttribute("Disciplina", 0, -5, 10, -5, (int)beatsCoelhoHora);
